requirements.txt
asgiref==3.8.1
Django==5.0.6
django-filter==24.2
pillow==10.3.0
sqlparse==0.5.0
tzdata==2024.1
